switch integration to unit for coverage
